Check safety properties instead of NP cycles
Turns out that checking for non-progress cycles doesn't use the given LTL property at all, and looks only for labels in the model. I was just lucky that the property I was checking against corresponded to a liveness property. So, I am back to checking never claims and getting short, finite counterexamples.