Run nuXmv BMC once for each loopback
Run bounded model checking once for every possible loopback value, weaken on every counterexample found, and pick the weakest interval. This reduces the uncertainty in picking a single loopback value that may not be suitable, or leaving it as '*' which can repeatedly give us finite counterexamples, with which we can't prove that no weakening exists (if this is the case in the model). Unfortunately I was not able to find a way to run BMC multiple times in one nuXmv invocation, so here we just run the whole executable once for each loopback.